requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
hugchat==0.4.8
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
urllib3==2.2.1
